Bug Fixed.
Bug Description: cache_data method was creating an array of data where 'url' was used as key for link. But it should be 'link' as the fill method provided by IData interface accepts its name as "link". It was causing problem. Link was vanished due to this on many RSS source.